requirements.txt
asgiref==3.6.0
Django==4.1.6
djangorestframework==3.14.0
psycopg2-binary
pytz==2022.7.1
sqlparse==0.4.3
